Create IntArch_ClayTools_modelsOnly.Rmd
change structure / logic to use functions. There is now a ModelBase function that has the "guts" of the model, the part that is common to all alternative models. Then, each of the three models are specified (shouldn't there be four? the simple simple one that just looks at number of similarities is missing). Then at the end, each model's function is called to produce the posterior.