My writing folder - with intro outline & relevant materials
Take a look at the outline of the intro that I just made - the first page of the .docx file in the main folder here (not in a subfolder). It may be useful context for your methods writing? Feedback welcome too.

That's the only thing that is current - the text in the body of the paper is a mishmash, I'll work on the intro throughout this week.